UserService tested

+++ src/main/java/entity/User.java
@@ -1,17 +1,17 @@
 package entity;
 
 public class User {
-    private int id;
+    private Long id;
     private String login;
     private String password;
 
-    public User(int id, String login, String password) {
+    public User(Long id, String login, String password) {
         this.id = id;
         this.login = login;
         this.password = password;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
